The RadialGauge and LinearGauge are no longer generic. Remove LinearGaugeFor and RadialGaugeFor methods. Closes #289. Closes #290

+++ wrappers/mvc/src/Kendo.Mvc/UI/Chart/Fluent/ChartPlotBandsBuilder.cs
@@ -37,7 +37,7 @@ namespace Kendo.Mvc.UI.Fluent
         /// %&gt;
         /// </code>
         /// </example>        
-        public ChartPlotBandsBuilder From(decimal from)
+        public ChartPlotBandsBuilder From(double from)
         {
             item.From = from;
             return this;
@@ -60,7 +60,7 @@ namespace Kendo.Mvc.UI.Fluent
         /// %&gt;
         /// </code>
         /// </example>        
-        public ChartPlotBandsBuilder To(decimal to)
+        public ChartPlotBandsBuilder To(double to)
         {
             item.To = to;
             return this;
